Fix: 에피그램 폼 리팩토링
